pyproject.toml
[project]
name = "swing-screener"
version = "0.1.0"
description = "Multi-source swing trading screener with pre-breakout and breakout detection"
requires-python = ">=3.10"
dependencies = [
    "finnhub-python>=2.4.20",
    "yfinance>=0.2.40",
    "pandas>=2.0",
    "numpy>=1.24",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "lxml>=4.9",
    "tabulate>=0.9",
    "pyarrow>=14.0",
]

[project.scripts]
swing-screener = "screener:main"
